backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
openai
pandas
plotly
requests
aiofiles
httpx
pytest
pytest-asyncio